+++ src/main/resources/gerenciadorfinanceiro.sql
@@ -39,7 +39,7 @@ create table if not exists tb_transacao_financeira (
 	identificador varchar(50) not null,
 	quantidade_parcela integer not null,
 	data_hora_cadastro timestamp not null, 
-	valor_total money not null,
+	valor_total float not null,
 	e_recorrente boolean not null,
 	e_pago boolean not null,
 	observacao varchar(255),
@@ -68,7 +68,7 @@ create table if not exists tb_parcelamento (
 	codigo bigserial not null,
 	identificador varchar(50) not null unique,
 	numero integer not null,
-	valor MONEY not null,
+	valor float not null,
 	data_vencimento date not null,
 	data_pagamento date not null,
 	constraint pk_parcelamento primary key (codigo)
